+++ static/index.html
@@ -6,9 +6,9 @@
   <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
   <link href="https://fonts.googleapis.com/css2?family=Cormorant:wght@300..700&family=Noto+Sans+TC:wght@100..900&family=Noto+Sans:ital,wght@0,100..900;1,100..900&family=Roboto:ital,wght@0,100;0,300;0,400;0,500;0,700;0,900;1,100;1,300;1,400;1,500;1,700;1,900&display=swap" rel="stylesheet">
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <script src="./static/script/script.js"></script>
-  <link rel="stylesheet" href="./static/style/style.css">
-  <link rel="stylesheet" href="./static/style/global.css">
+  <script type="module" src="/static/script/main.js" defer></script>
+  <link rel="stylesheet" href="/static/style/style.css">
+  <link rel="stylesheet" href="/static/style/global.css">
   <link href="https://cdn.jsdelivr.net/npm/@mdi/font/css/materialdesignicons.min.css" rel="stylesheet">
   <title>Taipei Day Trip</title>
 </head>
